create working DAOs and Model for groupbuy

+++ src/main/java/com/codecool/queststore/model/user/wallet/NoWallet.java
@@ -1,10 +1,8 @@
 package com.codecool.queststore.model.user.wallet;
 
-import com.codecool.queststore.model.user.User;
-
 public class NoWallet implements WalletStrategy{
     @Override
-    public Integer getWallet() {
+    public Double getWallet() {
         return null;
     }
 }
