+++ src/main/java/com/codecool/queststore/model/user/wallet/WalletStrategy.java
@@ -1,9 +1,7 @@
 package com.codecool.queststore.model.user.wallet;
 
-import com.codecool.queststore.model.user.User;
-
 public interface WalletStrategy {
 
-    Integer getWallet();
+    Double getWallet();
 
 }
